server/requirements.txt
asyncpg
fastapi
zeroconf2
uvicorn